.history/requirements_20250514091301.txt
# Development dependencies for testing and local scripts
pytest
websocket-client
